src/requirements.txt
scikit-learn==1.1.3
mlflow==2.0.1
pandas==1.5.2
numpy==1.23.4
pyyaml==6.0
seaborn==0.12.2
flask==2.2.3
dvc==2.52.0

